Moving jQuery to a peer dependency.
The present mainline way that jQuery is loaded seems to be wreaking
havoc on my Browserify + Babel pipeline.  There's a bunch of 'use
strict' that it's inserting.

The way I've usually solved the problem that your loading of $
is trying to solve, in the npm world, is to declare it as a
peer dependency.  And then require the user to load the $ part.